server/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
pydantic==1.10.7
python-dotenv==0.21.1
requests==2.28.2
jira==3.5.0
apscheduler==3.10.1
pytest==7.3.1
httpx==0.24.0
sqlalchemy==2.0.13